move mark-and-sweep gc to separate header
