+++ src/gc/mark_and_sweep_garbage_collection_scheme.cc
@@ -20,7 +20,7 @@ COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
 IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
 CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 *******************************************************************************/
-#include "../../include/gc/garbage_collection_scheme.h"
+#include "../../include/gc/mark_and_sweep_garbage_collection_scheme.h"
 
 #include <algorithm>
